+++ source/Core/Castor3D/Gui/Controls/CtrlSlider.cpp
@@ -34,7 +34,7 @@ namespace castor3d
 		, castor::RangedValue< int32_t > const & value
 		, castor::Position const & position
 		, castor::Size const & size
-		, uint32_t flags
+		, uint64_t flags
 		, bool visible )
 		: Control{ Type
 			, scene
